@@ -1,13 +1,14 @@
 {
   "name": "eslint-config-airbnb-base",
-  "version": "3.0.1",
+  "version": "9.0.0",
   "description": "Airbnb's base JS ESLint config, following our styleguide",
   "main": "index.js",
   "scripts": {
     "lint": "eslint .",
     "tests-only": "babel-tape-runner ./test/test-*.js",
-    "pretest": "eslint-find-rules --unused",
-    "test": "npm run --silent lint && npm run --silent tests-only",
+    "prepublish": "(in-install || eslint-find-rules --unused) && (not-in-publish || npm test) && safe-publish-latest",
+    "pretest": "npm run --silent lint",
+    "test": "npm run --silent tests-only",
     "travis": "npm run --silent test"
   },
   "repository": {
@@ -44,14 +45,20 @@
   },
   "homepage": "https://github.com/airbnb/javascript",
   "devDependencies": {
-    "babel-tape-runner": "^1.3.1",
-    "eslint": "^2.10.1",
-    "eslint-find-rules": "^1.9.2",
-    "eslint-plugin-import": "^1.8.0",
-    "tape": "^4.5.1"
+    "babel-preset-airbnb": "^2.1.1",
+    "babel-tape-runner": "^2.0.1",
+    "eslint": "^3.8.1",
+    "eslint-find-rules": "^1.14.0",
+    "eslint-plugin-import": "^2.0.1",
+    "in-publish": "^2.0.0",
+    "safe-publish-latest": "^1.1.1",
+    "tape": "^4.6.2"
   },
   "peerDependencies": {
-    "eslint": "^2.10.1",
-    "eslint-plugin-import": "^1.8.0"
+    "eslint": "^3.8.1",
+    "eslint-plugin-import": "^2.0.1"
+  },
+  "engines": {
+    "node": ">= 4"
   }
 }
