@@ -1,14 +1,15 @@
 {
   "name": "eslint-config-airbnb",
-  "version": "9.0.1",
+  "version": "12.0.0",
   "description": "Airbnb's ESLint config, following our styleguide",
   "main": "index.js",
   "scripts": {
     "lint": "eslint .",
     "tests-only": "babel-tape-runner ./test/test-*.js",
-    "pretest": "eslint-find-rules --unused",
-    "test": "npm run --silent lint && npm run --silent tests-only",
-    "travis": "cd ../eslint-config-airbnb-base && npm link && cd - && npm link eslint-config-airbnb-base && npm run --silent test ; npm unlink eslint-config-airbnb-base >/dev/null &"
+    "prepublish": "(in-install || eslint-find-rules --unused) && (not-in-publish || npm test) && safe-publish-latest",
+    "pretest": "npm run --silent lint",
+    "test": "npm run --silent tests-only",
+    "travis": "cd ../eslint-config-airbnb-base && npm install && npm link && cd - && npm link eslint-config-airbnb-base && npm run --silent test ; npm unlink eslint-config-airbnb-base >/dev/null &"
   },
   "repository": {
     "type": "git",
@@ -44,22 +45,28 @@
   },
   "homepage": "https://github.com/airbnb/javascript",
   "dependencies": {
-    "eslint-config-airbnb-base": "^3.0.1"
+    "eslint-config-airbnb-base": "^9.0.0"
   },
   "devDependencies": {
-    "babel-tape-runner": "^1.3.1",
-    "eslint": "^2.9.0",
-    "eslint-find-rules": "^1.9.2",
-    "eslint-plugin-import": "^1.8.0",
-    "eslint-plugin-jsx-a11y": "^1.2.0",
-    "eslint-plugin-react": "^5.1.1",
-    "react": "*",
-    "tape": "^4.5.1"
+    "babel-preset-airbnb": "^2.1.1",
+    "babel-tape-runner": "^2.0.1",
+    "eslint": "^3.8.1",
+    "eslint-find-rules": "^1.14.0",
+    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-react": "^6.4.1",
+    "in-publish": "^2.0.0",
+    "react": ">= 0.13.0",
+    "safe-publish-latest": "^1.1.1",
+    "tape": "^4.6.2"
   },
   "peerDependencies": {
-    "eslint": "^2.9.0",
-    "eslint-plugin-jsx-a11y": "^1.2.0",
-    "eslint-plugin-import": "^1.8.0",
-    "eslint-plugin-react": "^5.1.1"
+    "eslint": "^3.8.1",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-react": "^6.4.1"
+  },
+  "engines": {
+    "node": ">= 4"
   }
 }
