@@ -2,18 +2,21 @@ import fs from 'fs';
 import path from 'path';
 import test from 'tape';
 
-const index = require('../');
+import index from '../';
 
-const files = { index };
+const files = { ...{ index } }; // object spread is to test parsing
 
-fs.readdirSync(path.join(__dirname, '../rules')).forEach(name => {
+fs.readdirSync(path.join(__dirname, '../rules')).forEach((name) => {
+  // eslint-disable-next-line import/no-dynamic-require
   files[name] = require(`../rules/${name}`); // eslint-disable-line global-require
 });
 
-Object.keys(files).forEach(name => {
+Object.keys(files).forEach((
+  name, // trailing function comma is to test parsing
+) => {
   const config = files[name];
 
-  test(`${name}: does not reference react`, t => {
+  test(`${name}: does not reference react`, (t) => {
     t.plan(2);
 
     // scan plugins for react and fail if it is found
