@@ -1,14 +1,21 @@
 module.exports = {
   extends: [
-    './legacy',
+    './rules/best-practices',
+    './rules/errors',
+    './rules/node',
+    './rules/style',
+    './rules/variables',
     './rules/es6',
     './rules/imports',
   ].map(require.resolve),
   parserOptions: {
-    ecmaVersion: 7,
+    ecmaVersion: 2017,
     sourceType: 'module',
+    ecmaFeatures: {
+      experimentalObjectRestSpread: true,
+    },
   },
   rules: {
-    strict: 2,
-  }
+    strict: 'error',
+  },
 };
