@@ -1,60 +1,68 @@
 module.exports = {
-  'rules': {
+  rules: {
     // enforces getter/setter pairs in objects
-    'accessor-pairs': 0,
+    'accessor-pairs': 'off',
 
     // enforces return statements in callbacks of array's methods
     // http://eslint.org/docs/rules/array-callback-return
-    'array-callback-return': 2,
+    'array-callback-return': 'error',
 
     // treat var statements as if they were block scoped
-    'block-scoped-var': 2,
+    'block-scoped-var': 'error',
 
     // specify the maximum cyclomatic complexity allowed in a program
-    'complexity': [0, 11],
+    complexity: ['off', 11],
+
+    // enforce that class methods use "this"
+    // http://eslint.org/docs/rules/class-methods-use-this
+    'class-methods-use-this': ['error', {
+      exceptMethods: [],
+    }],
 
     // require return statements to either always or never specify values
-    'consistent-return': 2,
+    'consistent-return': 'error',
 
     // specify curly brace conventions for all control statements
-    'curly': [2, 'multi-line'],
+    curly: ['error', 'multi-line'],
 
     // require default case in switch statements
-    'default-case': [2, { 'commentPattern': '^no default$' }],
+    'default-case': ['error', { commentPattern: '^no default$' }],
 
     // encourages use of dot notation whenever possible
-    'dot-notation': [2, { 'allowKeywords': true }],
+    'dot-notation': ['error', { allowKeywords: true }],
 
     // enforces consistent newlines before or after dots
-    'dot-location': 0,
+    // http://eslint.org/docs/rules/dot-location
+    'dot-location': ['error', 'property'],
 
     // require the use of === and !==
     // http://eslint.org/docs/rules/eqeqeq
-    'eqeqeq': [2, 'allow-null'],
+    eqeqeq: ['error', 'allow-null'],
 
     // make sure for-in loops have an if statement
-    'guard-for-in': 2,
+    'guard-for-in': 'error',
 
     // disallow the use of alert, confirm, and prompt
-    'no-alert': 1,
+    'no-alert': 'warn',
 
     // disallow use of arguments.caller or arguments.callee
-    'no-caller': 2,
+    'no-caller': 'error',
 
     // disallow lexical declarations in case/default clauses
     // http://eslint.org/docs/rules/no-case-declarations.html
-    'no-case-declarations': 2,
+    'no-case-declarations': 'error',
 
     // disallow division operators explicitly at beginning of regular expression
-    'no-div-regex': 0,
+    // http://eslint.org/docs/rules/no-div-regex
+    'no-div-regex': 'off',
 
     // disallow else after a return in an if
-    'no-else-return': 2,
+    'no-else-return': 'error',
 
     // disallow empty functions, except for standalone funcs/arrows
     // http://eslint.org/docs/rules/no-empty-function
-    'no-empty-function': [2, {
-      'allow': [
+    'no-empty-function': ['error', {
+      allow: [
         'arrowFunctions',
         'functions',
         'methods',
@@ -63,161 +71,188 @@ module.exports = {
 
     // disallow empty destructuring patterns
     // http://eslint.org/docs/rules/no-empty-pattern
-    'no-empty-pattern': 2,
+    'no-empty-pattern': 'error',
 
     // disallow comparisons to null without a type-checking operator
-    'no-eq-null': 0,
+    'no-eq-null': 'off',
 
     // disallow use of eval()
-    'no-eval': 2,
+    'no-eval': 'error',
 
     // disallow adding to native types
-    'no-extend-native': 2,
+    'no-extend-native': 'error',
 
     // disallow unnecessary function binding
-    'no-extra-bind': 2,
+    'no-extra-bind': 'error',
 
     // disallow Unnecessary Labels
     // http://eslint.org/docs/rules/no-extra-label
-    'no-extra-label': 2,
+    'no-extra-label': 'error',
 
     // disallow fallthrough of case statements
-    'no-fallthrough': 2,
+    'no-fallthrough': 'error',
 
     // disallow the use of leading or trailing decimal points in numeric literals
-    'no-floating-decimal': 2,
+    'no-floating-decimal': 'error',
 
-    // disallow the type conversions with shorter notations
-    'no-implicit-coercion': 0,
+    // disallow reassignments of native objects or read-only globals
+    // http://eslint.org/docs/rules/no-global-assign
+    'no-global-assign': ['error', { exceptions: [] }],
+    // deprecated in favor of no-global-assign
+    'no-native-reassign': 'off',
+
+    // disallow implicit type conversions
+    // http://eslint.org/docs/rules/no-implicit-coercion
+    'no-implicit-coercion': ['off', {
+      boolean: false,
+      number: true,
+      string: true,
+      allow: [],
+    }],
 
     // disallow var and named functions in global scope
     // http://eslint.org/docs/rules/no-implicit-globals
-    'no-implicit-globals': 0,
+    'no-implicit-globals': 'off',
 
     // disallow use of eval()-like methods
-    'no-implied-eval': 2,
+    'no-implied-eval': 'error',
 
     // disallow this keywords outside of classes or class-like objects
-    'no-invalid-this': 0,
+    'no-invalid-this': 'off',
 
     // disallow usage of __iterator__ property
-    'no-iterator': 2,
+    'no-iterator': 'error',
 
     // disallow use of labels for anything other then loops and switches
-    'no-labels': [2, { 'allowLoop': false, 'allowSwitch': false }],
+    'no-labels': ['error', { allowLoop: false, allowSwitch: false }],
 
     // disallow unnecessary nested blocks
-    'no-lone-blocks': 2,
+    'no-lone-blocks': 'error',
 
     // disallow creation of functions within loops
-    'no-loop-func': 2,
+    'no-loop-func': 'error',
 
     // disallow magic numbers
     // http://eslint.org/docs/rules/no-magic-numbers
-    'no-magic-numbers': [0, {
-      'ignore': [],
-      'ignoreArrayIndexes': true,
-      'enforceConst': true,
-      'detectObjects': false,
+    'no-magic-numbers': ['off', {
+      ignore: [],
+      ignoreArrayIndexes: true,
+      enforceConst: true,
+      detectObjects: false,
     }],
 
     // disallow use of multiple spaces
-    'no-multi-spaces': 2,
+    'no-multi-spaces': 'error',
 
     // disallow use of multiline strings
-    'no-multi-str': 2,
-
-    // disallow reassignments of native objects
-    'no-native-reassign': 2,
+    'no-multi-str': 'error',
 
     // disallow use of new operator when not part of the assignment or comparison
-    'no-new': 2,
+    'no-new': 'error',
 
     // disallow use of new operator for Function object
-    'no-new-func': 2,
+    'no-new-func': 'error',
 
     // disallows creating new instances of String, Number, and Boolean
-    'no-new-wrappers': 2,
+    'no-new-wrappers': 'error',
 
     // disallow use of (old style) octal literals
-    'no-octal': 2,
+    'no-octal': 'error',
 
     // disallow use of octal escape sequences in string literals, such as
     // var foo = 'Copyright \251';
-    'no-octal-escape': 2,
+    'no-octal-escape': 'error',
 
     // disallow reassignment of function parameters
     // disallow parameter object manipulation
     // rule: http://eslint.org/docs/rules/no-param-reassign.html
-    'no-param-reassign': [2, { 'props': true }],
+    'no-param-reassign': ['error', { props: true }],
 
     // disallow usage of __proto__ property
-    'no-proto': 2,
+    'no-proto': 'error',
 
     // disallow declaring the same variable more then once
-    'no-redeclare': 2,
+    'no-redeclare': 'error',
+
+    // disallow certain object properties
+    // http://eslint.org/docs/rules/no-restricted-properties
+    'no-restricted-properties': ['error', {
+      object: 'arguments',
+      property: 'callee',
+      message: 'arguments.callee is deprecated,'
+    }, {
+      property: '__defineGetter__',
+      message: 'Please use Object.defineProperty instead.',
+    }, {
+      property: '__defineSetter__',
+      message: 'Please use Object.defineProperty instead.',
+    }],
 
     // disallow use of assignment in return statement
-    'no-return-assign': 2,
+    'no-return-assign': 'error',
 
     // disallow use of `javascript:` urls.
-    'no-script-url': 2,
+    'no-script-url': 'error',
 
     // disallow self assignment
     // http://eslint.org/docs/rules/no-self-assign
-    'no-self-assign': 2,
+    'no-self-assign': 'error',
 
     // disallow comparisons where both sides are exactly the same
-    'no-self-compare': 2,
+    'no-self-compare': 'error',
 
     // disallow use of comma operator
-    'no-sequences': 2,
+    'no-sequences': 'error',
 
     // restrict what can be thrown as an exception
-    'no-throw-literal': 2,
+    'no-throw-literal': 'error',
 
     // disallow unmodified conditions of loops
     // http://eslint.org/docs/rules/no-unmodified-loop-condition
-    'no-unmodified-loop-condition': 0,
+    'no-unmodified-loop-condition': 'off',
 
     // disallow usage of expressions in statement position
-    'no-unused-expressions': 2,
+    'no-unused-expressions': ['error', {
+      allowShortCircuit: false,
+      allowTernary: false,
+    }],
 
     // disallow unused labels
     // http://eslint.org/docs/rules/no-unused-labels
-    'no-unused-labels': 2,
+    'no-unused-labels': 'error',
 
     // disallow unnecessary .call() and .apply()
-    'no-useless-call': 0,
+    'no-useless-call': 'off',
 
     // disallow useless string concatenation
     // http://eslint.org/docs/rules/no-useless-concat
-    'no-useless-concat': 2,
+    'no-useless-concat': 'error',
 
     // disallow unnecessary string escaping
     // http://eslint.org/docs/rules/no-useless-escape
-    'no-useless-escape': 2,
+    'no-useless-escape': 'error',
 
     // disallow use of void operator
-    'no-void': 0,
+    // http://eslint.org/docs/rules/no-void
+    'no-void': 'error',
 
     // disallow usage of configurable warning terms in comments: e.g. todo
-    'no-warning-comments': [0, { 'terms': ['todo', 'fixme', 'xxx'], 'location': 'start' }],
+    'no-warning-comments': ['off', { terms: ['todo', 'fixme', 'xxx'], location: 'start' }],
 
     // disallow use of the with statement
-    'no-with': 2,
+    'no-with': 'error',
 
     // require use of the second argument for parseInt()
-    'radix': 2,
+    radix: 'error',
 
     // requires to declare all vars on top of their containing scope
-    'vars-on-top': 2,
+    'vars-on-top': 'error',
 
     // require immediate function invocation to be wrapped in parentheses
     // http://eslint.org/docs/rules/wrap-iife.html
-    'wrap-iife': [2, 'outside'],
+    'wrap-iife': ['error', 'outside', { functionPrototypeMethods: false }],
 
     // require or disallow Yoda conditions
-    'yoda': 2
+    yoda: 'error'
   }
 };
