@@ -1,124 +1,159 @@
 module.exports = {
-  'env': {
-    'es6': true
+  env: {
+    es6: true
   },
-  'parserOptions': {
-    'ecmaVersion': 6,
-    'sourceType': 'module',
-    'ecmaFeatures': {
-      'generators': false,
-      'objectLiteralDuplicateProperties': false
+  parserOptions: {
+    ecmaVersion: 6,
+    sourceType: 'module',
+    ecmaFeatures: {
+      generators: false,
+      objectLiteralDuplicateProperties: false
     }
   },
 
-  'rules': {
+  rules: {
     // enforces no braces where they can be omitted
     // http://eslint.org/docs/rules/arrow-body-style
-    'arrow-body-style': [2, 'as-needed'],
+    // TODO: enable requireReturnForObjectLiteral?
+    'arrow-body-style': ['error', 'as-needed', {
+      requireReturnForObjectLiteral: false,
+    }],
 
     // require parens in arrow function arguments
-    'arrow-parens': 0,
+    // http://eslint.org/docs/rules/arrow-parens
+    'arrow-parens': ['error', 'as-needed', {
+      requireForBlockBody: true,
+    }],
 
     // require space before/after arrow function's arrow
     // http://eslint.org/docs/rules/arrow-spacing
-    'arrow-spacing': [2, { 'before': true, 'after': true }],
+    'arrow-spacing': ['error', { before: true, after: true }],
 
     // verify super() callings in constructors
-    'constructor-super': 0,
+    'constructor-super': 'error',
 
     // enforce the spacing around the * in generator functions
     // http://eslint.org/docs/rules/generator-star-spacing
-    'generator-star-spacing': [2, { 'before': false, 'after': true }],
+    'generator-star-spacing': ['error', { before: false, after: true }],
 
     // disallow modifying variables of class declarations
     // http://eslint.org/docs/rules/no-class-assign
-    'no-class-assign': 2,
+    'no-class-assign': 'error',
 
     // disallow arrow functions where they could be confused with comparisons
     // http://eslint.org/docs/rules/no-confusing-arrow
-    'no-confusing-arrow': [2, {
-      'allowParens': true,
+    'no-confusing-arrow': ['error', {
+      allowParens: true,
     }],
 
     // disallow modifying variables that are declared using const
-    'no-const-assign': 2,
+    'no-const-assign': 'error',
 
     // disallow duplicate class members
     // http://eslint.org/docs/rules/no-dupe-class-members
-    'no-dupe-class-members': 2,
+    'no-dupe-class-members': 'error',
 
     // disallow importing from the same path more than once
     // http://eslint.org/docs/rules/no-duplicate-imports
-    'no-duplicate-imports': 2,
+    'no-duplicate-imports': 'error',
 
     // disallow symbol constructor
     // http://eslint.org/docs/rules/no-new-symbol
-    'no-new-symbol': 2,
+    'no-new-symbol': 'error',
 
     // disallow specific imports
     // http://eslint.org/docs/rules/no-restricted-imports
-    'no-restricted-imports': 0,
+    'no-restricted-imports': 'off',
 
     // disallow to use this/super before super() calling in constructors.
-    'no-this-before-super': 0,
+    // http://eslint.org/docs/rules/no-this-before-super
+    'no-this-before-super': 'error',
 
     // disallow useless computed property keys
     // http://eslint.org/docs/rules/no-useless-computed-key
-    'no-useless-computed-key': 2,
+    'no-useless-computed-key': 'error',
 
     // disallow unnecessary constructor
     // http://eslint.org/docs/rules/no-useless-constructor
-    'no-useless-constructor': 2,
+    'no-useless-constructor': 'error',
+
+    // disallow renaming import, export, and destructured assignments to the same name
+    // http://eslint.org/docs/rules/no-useless-rename
+    'no-useless-rename': ['error', {
+      ignoreDestructuring: false,
+      ignoreImport: false,
+      ignoreExport: false,
+    }],
 
     // require let or const instead of var
-    'no-var': 2,
+    'no-var': 'error',
 
     // require method and property shorthand syntax for object literals
     // http://eslint.org/docs/rules/object-shorthand
-    'object-shorthand': [2, 'always', {
-      'ignoreConstructors': false,
-      'avoidQuotes': true,
+    'object-shorthand': ['error', 'always', {
+      ignoreConstructors: false,
+      avoidQuotes: true,
     }],
 
     // suggest using arrow functions as callbacks
-    'prefer-arrow-callback': [2, {
-      'allowNamedFunctions': false,
-      'allowUnboundThis': true,
+    'prefer-arrow-callback': ['error', {
+      allowNamedFunctions: false,
+      allowUnboundThis: true,
     }],
 
     // suggest using of const declaration for variables that are never modified after declared
-    'prefer-const': [2, {
-      'destructuring': 'any',
-      'ignoreReadBeforeAssign': true,
+    'prefer-const': ['error', {
+      destructuring: 'any',
+      ignoreReadBeforeAssign: true,
     }],
 
+    // disallow parseInt() in favor of binary, octal, and hexadecimal literals
+    // http://eslint.org/docs/rules/prefer-numeric-literals
+    'prefer-numeric-literals': 'error',
+
     // suggest using Reflect methods where applicable
-    'prefer-reflect': 0,
+    // http://eslint.org/docs/rules/prefer-reflect
+    // TODO: enable?
+    'prefer-reflect': 'off',
 
     // use rest parameters instead of arguments
     // http://eslint.org/docs/rules/prefer-rest-params
-    'prefer-rest-params': 2,
+    'prefer-rest-params': 'error',
 
     // suggest using the spread operator instead of .apply()
-    'prefer-spread': 0,
+    // http://eslint.org/docs/rules/prefer-spread
+    'prefer-spread': 'error',
 
     // suggest using template literals instead of string concatenation
     // http://eslint.org/docs/rules/prefer-template
-    'prefer-template': 2,
+    'prefer-template': 'error',
 
     // disallow generator functions that do not have yield
-    'require-yield': 0,
+    // http://eslint.org/docs/rules/require-yield
+    'require-yield': 'error',
+
+    // enforce spacing between object rest-spread
+    // http://eslint.org/docs/rules/rest-spread-spacing
+    'rest-spread-spacing': ['error', 'never'],
 
     // import sorting
     // http://eslint.org/docs/rules/sort-imports
-    'sort-imports': 0,
+    'sort-imports': ['off', {
+      ignoreCase: false,
+      ignoreMemberSort: false,
+      memberSyntaxSortOrder: ['none', 'all', 'multiple', 'single'],
+    }],
+
+    // require a Symbol description
+    // http://eslint.org/docs/rules/symbol-description
+    'symbol-description': 'error',
 
     // enforce usage of spacing in template strings
     // http://eslint.org/docs/rules/template-curly-spacing
-    'template-curly-spacing': 2,
+    'template-curly-spacing': 'error',
 
     // enforce spacing around the * in yield* expressions
     // http://eslint.org/docs/rules/yield-star-spacing
-    'yield-star-spacing': [2, 'after']
+    'yield-star-spacing': ['error', 'after']
   }
 };
