@@ -1,38 +1,39 @@
 module.exports = {
-  'env': {
-    'node': true
+  env: {
+    node: true
   },
 
-  'rules': {
+  rules: {
     // enforce return after a callback
-    'callback-return': 0,
+    'callback-return': 'off',
 
     // require all requires be top-level
     // http://eslint.org/docs/rules/global-require
-    'global-require': 2,
+    'global-require': 'error',
 
     // enforces error handling in callbacks (node environment)
-    'handle-callback-err': 0,
+    'handle-callback-err': 'off',
 
     // disallow mixing regular variable and require declarations
-    'no-mixed-requires': [0, false],
+    'no-mixed-requires': ['off', false],
 
     // disallow use of new operator with the require function
-    'no-new-require': 0,
+    'no-new-require': 'error',
 
     // disallow string concatenation with __dirname and __filename
-    'no-path-concat': 0,
+    // http://eslint.org/docs/rules/no-path-concat
+    'no-path-concat': 'error',
 
     // disallow use of process.env
-    'no-process-env': 0,
+    'no-process-env': 'off',
 
     // disallow process.exit()
-    'no-process-exit': 0,
+    'no-process-exit': 'off',
 
     // restrict usage of specified node modules
-    'no-restricted-modules': 0,
+    'no-restricted-modules': 'off',
 
     // disallow use of synchronous methods (off by default)
-    'no-sync': 0
+    'no-sync': 'off',
   }
 };
