@@ -1,6 +1,6 @@
 module.exports = {
-  'rules': {
+  rules: {
     // babel inserts `'use strict';` for us
-    'strict': [2, 'never']
+    strict: ['error', 'never']
   }
 };
