@@ -1,39 +1,42 @@
 module.exports = {
-  'rules': {
+  rules: {
     // enforce or disallow variable initializations at definition
-    'init-declarations': 0,
+    'init-declarations': 'off',
 
     // disallow the catch clause parameter name being the same as a variable in the outer scope
-    'no-catch-shadow': 0,
+    'no-catch-shadow': 'off',
 
     // disallow deletion of variables
-    'no-delete-var': 2,
+    'no-delete-var': 'error',
 
     // disallow labels that share a name with a variable
-    'no-label-var': 0,
+    // http://eslint.org/docs/rules/no-label-var
+    'no-label-var': 'error',
 
     // disallow specific globals
-    'no-restricted-globals': 0,
+    'no-restricted-globals': 'off',
 
     // disallow declaration of variables already declared in the outer scope
-    'no-shadow': 2,
+    'no-shadow': 'error',
 
     // disallow shadowing of names such as arguments
-    'no-shadow-restricted-names': 2,
+    'no-shadow-restricted-names': 'error',
 
     // disallow use of undeclared variables unless mentioned in a /*global */ block
-    'no-undef': 2,
+    'no-undef': 'error',
 
     // disallow use of undefined when initializing variables
-    'no-undef-init': 0,
+    'no-undef-init': 'error',
 
     // disallow use of undefined variable
-    'no-undefined': 0,
+    // http://eslint.org/docs/rules/no-undefined
+    // TODO: enable?
+    'no-undefined': 'off',
 
     // disallow declaration of variables that are not used in the code
-    'no-unused-vars': [2, { 'vars': 'local', 'args': 'after-used' }],
+    'no-unused-vars': ['error', { vars: 'local', args: 'after-used' }],
 
     // disallow use of variables before they are defined
-    'no-use-before-define': 2
+    'no-use-before-define': 'error'
   }
 };
