@@ -6,18 +6,19 @@ const base = require('../base');
 
 const files = { base };
 
-fs.readdirSync(path.join(__dirname, '../rules')).forEach(name => {
+fs.readdirSync(path.join(__dirname, '../rules')).forEach((name) => {
   if (name === 'react.js' || name === 'react-a11y.js') {
     return;
   }
 
+  // eslint-disable-next-line import/no-dynamic-require
   files[name] = require(`../rules/${name}`); // eslint-disable-line global-require
 });
 
-Object.keys(files).forEach(name => {
+Object.keys(files).forEach((name) => {
   const config = files[name];
 
-  test(`${name}: does not reference react`, t => {
+  test(`${name}: does not reference react`, (t) => {
     t.plan(2);
 
     // scan plugins for react and fail if it is found
