@@ -8,8 +8,10 @@ const cli = new CLIEngine({
   useEslintrc: false,
   baseConfig: eslintrc,
 
-  // This rule fails when executing on text.
-  rules: { indent: 0 },
+  rules: {
+    // It is okay to import devDependencies in tests.
+    'import/no-extraneous-dependencies': [2, { devDependencies: true }],
+  },
 });
 
 function lint(text) {
@@ -22,8 +24,9 @@ function lint(text) {
 function wrapComponent(body) {
   return `
 import React from 'react';
+
 export default class MyComponent extends React.Component {
-/* eslint no-empty-function: 0 */
+/* eslint no-empty-function: 0, class-methods-use-this: 0 */
 ${body}
 }
 `;
@@ -54,7 +57,7 @@ test('validate react prop order', (t) => {
     t.deepEquals(result.messages, [], 'no messages in results');
   });
 
-  t.test('order: when random method is first', t => {
+  t.test('order: when random method is first', (t) => {
     t.plan(2);
     const result = lint(wrapComponent(`
   someMethod() {}
@@ -71,7 +74,7 @@ test('validate react prop order', (t) => {
     t.equal(result.messages[0].ruleId, 'react/sort-comp', 'fails due to sort');
   });
 
-  t.test('order: when random method after lifecycle methods', t => {
+  t.test('order: when random method after lifecycle methods', (t) => {
     t.plan(2);
     const result = lint(wrapComponent(`
   componentWillMount() {}
